Create linked list
Linked list class keeps track of all the nodes
with the head and tail properties. Add a push
method in order to add a new node to the end of
the list.

Node class is only responsible for creating a new
node: data and a pointer to the next node.